fastapi/requirements.txt
fastapi==0.110.0
pydantic==1.10.13
uvicorn==0.18.3
python-multipart==0.0.9
psycopg2-binary==2.9.9
langchain==0.2.5
langchain-openai==0.1.8
chromadb==0.4.24
markdownify==0.12.1
pypdf==4.2.0
sentence-transformers==2.7.0
ipython==8.18.1
pyautogen[retrievechat]==0.2.27
pgvector==0.2.5
psycopg==3.1.18
psycopg[binary,pool]
PyPDF2==3.0.1
langchain_experimental==0.0.61
pytest==8.2.2